Add files via upload
This is my Jupyter Notebook for Project Benson.  It contains weekly MTA turnstile data for all of 2018 and the first week of 2019.  Weekly files are in the format "MTA_mm_dd_yyR".  
The notebook also contains monthly files in the format "MTA_mm_2018R".  These files are rolled up into quarterly files in the form of "MTA_Qx_2018R", semiannual files in the form of "MTA_Hx_2018R".  The file with all the data for 2018 is called "MTA_2018R".

 All the files end with "R", which stands for "raw data".  No data in the original text files has been altered.   All I have done is to convert the original text files to pandas dataframes.  Please DO NOT make any inline changes to these files.  Any transformations should be made to COPIES of these dataframes.  That way, we have access to the original data if somehow a mistake is made along the way.  Believe me, it happens!